generated_projects/local_discord_bot/requirements.txt
discord.py[voice]
python-dotenv
pytest
